Replace certbot-auto with certbot
Since certbot-auto has reached the EOL, it is needed to replace this
method with on that actuall works. In this case, certbot has a method
that uses pip to install certbot and only needs some minor tweaks.

When building a new instance, the certman command would not work
because the python interpretor was wrong. Upon further investigation,
I noticed that there were some diffrences between what is on the server
now and what was in this repo. Neither of these changes expose the
environment and would be better changes here then in our configuration
manager.

Relates to: [BGO-580]