Merge remote server changes in Seth's repo dir
There appears to be some changes that were not pushed out to the repo,
but instead were left on the server. This caused an issue when trying to
run an update via ansible since the files do not match. I went to the
server that has a working version of certman with the changes that have
been working so far and copied it to my local computer. With that
version, I ran a diff between the BG repo and Seth's working server
version. What ever differences were merged over to the BG repo so that
we will continue to have them. The only exception was to the requests
and pyaml library versions listed in the requirements file since they
were older then what was in our repo.

Relates to: [BGO-580]